+++ client/client.go
@@ -10,11 +10,11 @@ import (
 
 	"github.com/Lagrange-Labs/lagrange-node/core"
 	"github.com/Lagrange-Labs/lagrange-node/core/logger"
+	"github.com/Lagrange-Labs/lagrange-node/core/telemetry"
 	"github.com/Lagrange-Labs/lagrange-node/rpcclient"
 	sequencerv2types "github.com/Lagrange-Labs/lagrange-node/sequencer/types/v2"
 	"github.com/Lagrange-Labs/lagrange-node/server"
 	serverv2types "github.com/Lagrange-Labs/lagrange-node/server/types/v2"
-	"github.com/Lagrange-Labs/lagrange-node/telemetry"
 )
 
 var (
