+++ client/health.go
@@ -11,8 +11,8 @@ import (
 
 	"github.com/Lagrange-Labs/lagrange-node/core"
 	"github.com/Lagrange-Labs/lagrange-node/core/logger"
+	"github.com/Lagrange-Labs/lagrange-node/core/telemetry"
 	serverv2types "github.com/Lagrange-Labs/lagrange-node/server/types/v2"
-	"github.com/Lagrange-Labs/lagrange-node/telemetry"
 )
 
 const (
