+++ server/service.go
@@ -13,9 +13,9 @@ import (
 	"github.com/Lagrange-Labs/lagrange-node/core"
 	"github.com/Lagrange-Labs/lagrange-node/core/crypto"
 	"github.com/Lagrange-Labs/lagrange-node/core/logger"
+	"github.com/Lagrange-Labs/lagrange-node/core/telemetry"
 	"github.com/Lagrange-Labs/lagrange-node/server/types"
 	v2types "github.com/Lagrange-Labs/lagrange-node/server/types/v2"
-	"github.com/Lagrange-Labs/lagrange-node/telemetry"
 )
 
 var (
